requirements.txt
sentence-transformers
faiss-cpu
spacy
openai
streamlit